change service from attribute to subcollection, finish add and edit service in room detail, delete service

+++ app/src/main/java/com/example/baytro/view/components/ServiceCard.kt
@@ -55,8 +55,8 @@ fun ServiceActionButton(
 @Composable
 fun ServiceCard(
     service: Service,
-    onEdit: (Service) -> Unit,
-    onDelete: (Service) -> Unit
+    onEdit: ((Service) -> Unit)?,
+    onDelete: ((Service) -> Unit)?
 ) {
     Box(
         modifier = Modifier
@@ -103,7 +103,7 @@ fun ServiceCard(
                     verticalAlignment = Alignment.CenterVertically,
                     horizontalArrangement = Arrangement.End
                 ) {
-                    if(onEdit(service) != {}) {
+                    if(onEdit != null) {
                         ServiceActionButton(
                             icon = Icons.Filled.Edit,
                             contentDescription = "Edit",
@@ -115,18 +115,17 @@ fun ServiceCard(
                         modifier = Modifier
                             .width(16.dp)
                     )
-
-                    if (onDelete(service) != {}) {
+                    // Chỉ hiển thị nút xóa cho các dịch vụ không phải là nước và điện
+                    if(
+                        onDelete != null &&
+                        !service.name.equals("water", ignoreCase = true) &&
+                        !service.name.equals("electricity", ignoreCase = true)
+                    ) {
                         ServiceActionButton(
                             icon = Icons.Filled.Delete,
                             contentDescription = "Delete",
                             backgroundColor = MaterialTheme.colorScheme.secondaryContainer,
                             onClick = { onDelete(service) }
-                        )
-
-                        Spacer(
-                            modifier = Modifier
-                                .width(16.dp)
                         )
                     }
                 }
